Add i18n support for translation support
Edited the look and feel of the settings to make it look like the room booking selections. Translations will need to be worked on, but the foundation is there now. Also created a spinner for loading the translation, as the translations will be loaded lazily.

+++ OfficeCalendar.API/OfficeCalendar.API/Services/SettingsService.cs
@@ -49,7 +49,7 @@ public class SettingsService : ISettingsService
             {
                 SiteTheme = settings.SiteTheme,
                 AccentColor = settings.AccentColor,
-                FontSize = settings.FontSize,
+                FontSize = (int)settings.FontSize,
                 DefaultCalendarView = settings.DefaultCalendarView,
                 Language = settings.Language
             };
